backend/requirements.txt
fastapi>=0.95.0
pydantic>=1.10.7
uvicorn>=0.21.1
python-multipart>=0.0.5
torch>=2.0.0
transformers>=4.28.0
diffusers>=0.16.0
pillow>=9.5.0
huggingface-hub>=0.14.1